requirements.txt
Django~=3.1.6
djangorestframework~=3.12.4
pyyaml
uritemplate
gunicorn
whitenoise
django-rest-swagger
drf_yasg
PyJWT~=2.0.1
djangorestframework-jwt